Added interactor list to resource script changed interactor code.
The other simpler method wasn't working for resources limited to 1 maxInteractor. They never made it to the farms, they just stand there toggling the interactors variable on the farm. The new system works, but not perfectly still. The problem with the old system was their goal was changing too often but it was changing to farming each time.